backend/requirements.txt
Django==2.0.4
django-rest-framework==0.1.0
djangorestframework==3.8.2
pkg-resources==0.0.0
pytz==2018.4
